\name{BDP2-package}
\alias{BDP2-package}
\docType{package}
\title{
\packageTitle{BDP2}
}
\description{
\packageDescription{BDP2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BDP2}
\packageIndices{BDP2}
Workflow to choose design parameters single-arm trial with a binary
endpoint (response, success) and interim efficacy and futility analyses as well as routines to determine and visualize operating characteristics.

Main functions: BDP2workflow(), BDP2() and plotBDP2().
}
\author{
\packageAuthor{BDP2}

Maintainer: \packageMaintainer{BDP2}
}
\references{
%%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
